requirements.txt
qdrant-client
sentence-transformers
faiss-cpu